Update tests for services
