+++ tests/test_core_fetchers.py
