+++ tests/test_core_summarizers.py
